refactor: move unhandled error listeners into separate file in renderer process
