+++ packages/renderer-process/src/rendererProcessMain.js
@@ -1,19 +1,19 @@
 import * as Command from './parts/Command/Command.js'
-import * as ErrorHandling from './parts/ErrorHandling/ErrorHandling.js'
 import * as Event from './parts/Event/Event.js'
 import * as HandleContentSecurityPolicyViolation from './parts/HandleContentSecurityPolicyViolation/HandleContentSecurityPolicyViolation.js'
 import * as Module from './parts/Module/Module.js'
 import * as Platform from './parts/Platform/Platform.js'
 import * as PlatformType from './parts/PlatformType/PlatformType.js'
 import * as RendererWorker from './parts/RendererWorker/RendererWorker.js'
+import * as UnhandledErrorHandling from './parts/UnhandledErrorHandling/UnhandledErrorHandling.js'
 
 const handleBeforeInstallPrompt = (event) => {
   Event.preventDefault(event)
 }
 
 const main = async () => {
-  onerror = ErrorHandling.handleError
-  onunhandledrejection = ErrorHandling.handleUnhandledRejection
+  onerror = UnhandledErrorHandling.handleUnhandledError
+  onunhandledrejection = UnhandledErrorHandling.handleUnhandledRejection
   if ('SecurityPolicyViolationEvent' in self) {
     self.addEventListener('securitypolicyviolation', HandleContentSecurityPolicyViolation.handleContentSecurityPolicyViolation)
   }
