+++ packages/renderer-process/test/UnhandledErrorHandling.test.js
@@ -12,25 +12,25 @@ jest.unstable_mockModule('../src/parts/Logger/Logger.js', () => {
   }
 })
 
-const ErrorHandling = await import('../src/parts/ErrorHandling/ErrorHandling.js')
+const UnhandledErrorHandling = await import('../src/parts/UnhandledErrorHandling/UnhandledErrorHandling.js')
 const Logger = await import('../src/parts/Logger/Logger.js')
 
 test('handleError - normal error', () => {
   const mockError = new Error('oops')
-  ErrorHandling.handleError(mockError)
+  UnhandledErrorHandling.handleUnhandledError(mockError)
   expect(Logger.info).toHaveBeenCalledTimes(1)
   expect(Logger.info).toHaveBeenCalledWith(`[renderer-process] Unhandled Error: Error: oops`)
 })
 
 test('handleError - null', () => {
   const mockError = null
-  ErrorHandling.handleError(mockError)
+  UnhandledErrorHandling.handleUnhandledError(mockError)
   expect(Logger.info).toHaveBeenCalledTimes(1)
   expect(Logger.info).toHaveBeenCalledWith(`[renderer-process] Unhandled Error: null`)
 })
 
 test('handleError - firefox worker error', () => {
   const mockError = `SyntaxError: import declarations may only appear at top level of a module`
-  ErrorHandling.handleError(mockError)
+  UnhandledErrorHandling.handleUnhandledError(mockError)
   expect(Logger.info).not.toHaveBeenCalled()
 })
